Back/requirements.txt
fastapi >= 0.112.2
fastapi[standard] >= 0.112.2
uvicorn >= 0.30.6
cryptography != 3.4.0, >= 3.1
pyjwt >= 1.3.1
bcrypt >= 4.2.0
psycopg2-binary >= 2.9.9
requests >= 2.32.3
logging >= 0.4.9.6
httpx >= 0.27.2
python-multipart >= 0.0.9
